fs
The core topic of the paper is concerned with the prediction of short-term market direction movement for bitcoin spot price exploring traditional statistics, machine learning and deep learning methods applied to time series data. We consider this a critical step as it forms the core of an algorithmic trading strategy that could be then deployed in real-time on the cloud. Here, the basic idea for modeling future market direction of a single asset, is based on predicting bitcoin's sign of future returns and transforming a sequence; in this case, time series-data, into a matrix structure where we have features and labels. We select the 25 most important features with SelectKBest as we include technical analysis indicators and on-chain data metrics as features to the mix for our supervised learning algorithms. We let them "learn" about the relationships between features and labels data from a simple classification problem approach. Finally, based on the premise "that a process of directional price changes is predictable if the accuracy of the predictions is significantly higher than 50%", we train and back-test a strategy in bitcoin based on our algorithms for directional (long/short) trading and compare the different models strategies perfomances' numerically and visually.